Add share action to NetworkService
Now the share action is in a totally different service, it means that it no longer bother the user
when the sharing is ok. Also it will not block the interface.

Also it means that a notification is sent when the share fail, the notification allow the user to
try again his share. For the moment there is one notification for each link, in the future I should
create single notification for all failed shares.

TODO :
 - Allow the user to send me an email if the share fails too many times.
 - Extend the layout of the notification for a better interface
 - Move other smaller network actions to the NetworkService